refactor fare product title
